Refactor code to make init less blob-like.
I started adding a few new files to the program to try to decouple
everything. init will call createMenu, which will take in parameters
from the functions in titleInit and optionsInit called in Init. I also
figured out a major problem we were having. We weren't compiling in
C++11, so that's where a lot of our errors were coming in from the std
functions and auto i. I added the flag to the makefile to fix it.